requirements.txt
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pdfplumber==0.11.4
pandas==2.2.3
python-dotenv==1.0.1
schedule==1.2.2
requests==2.32.3
python-dateutil==2.9.0